refactor: Clean-up shared state for NewPost
