+++ src/components/NewPost.js
@@ -1,14 +1,36 @@
-import { useContext } from 'react';
+import { useState, useContext } from 'react';
+import { useHistory } from 'react-router-dom';
+import api from '../api/posts';
 import { DataContext } from '../context/DataContext';
 
 const NewPost = () => {
-  const {
-    handleSubmit,
-    postTitle,
-    setPostTitle,
-    postBody,
-    setPostBody,
-  } = useContext(DataContext);
+  const { posts, setPosts, format } = useContext(DataContext);
+  const [postTitle, setPostTitle] = useState('');
+  const [postBody, setPostBody] = useState('');
+
+  const history = useHistory();
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    const id = posts.length ? posts[posts.length - 1].id + 1 : 1;
+    const datetime = format(new Date(), 'MMMM dd, yyyy HH:mm:ss');
+    const newPost = {
+      id,
+      title: postTitle,
+      body: postBody,
+      datetime: datetime,
+    };
+    try {
+      const response = await api.post('/posts', newPost);
+      const allPosts = [...posts, response.data];
+      setPosts(allPosts);
+      setPostTitle('');
+      setPostBody('');
+      history.push('/');
+    } catch (error) {
+      console.log(`Something went wrong: ${error.message}`);
+    }
+  };
 
   return (
     <main className='NewPost'>
